Move struct `syntax` into header `kernel/syntax.hpp`
